# pkg/query: Don't use GetValueIndex, use raw underlying array