--- a/pkg/profile/reader.go
+++ b/pkg/profile/reader.go
@@ -36,26 +36,26 @@ type RecordReader struct {
 	LabelFields  []arrow.Field
 	LabelColumns []LabelColumn
 
-	Locations                  *array.List
-	Location                   *array.Struct
-	Address                    *array.Uint64
-	MappingStart               *array.Uint64
-	MappingLimit               *array.Uint64
-	MappingOffset              *array.Uint64
-	MappingFile                *array.Dictionary
-	MappingFileDict            *array.Binary
-	MappingBuildID             *array.Dictionary
-	MappingBuildIDDict         *array.Binary
-	Lines                      *array.List
-	Line                       *array.Struct
-	LineNumber                 *array.Int64
-	LineFunctionName           *array.Dictionary
-	LineFunctionNameDict       *array.Binary
-	LineFunctionSystemName     *array.Dictionary
-	LineFunctionSystemNameDict *array.Binary
-	LineFunctionFilename       *array.Dictionary
-	LineFunctionFilenameDict   *array.Binary
-	LineFunctionStartLine      *array.Int64
+	Locations                     *array.List
+	Location                      *array.Struct
+	Address                       *array.Uint64
+	MappingStart                  *array.Uint64
+	MappingLimit                  *array.Uint64
+	MappingOffset                 *array.Uint64
+	MappingFileIndices            *array.Uint32
+	MappingFileDict               *array.Binary
+	MappingBuildIDIndices         *array.Uint32
+	MappingBuildIDDict            *array.Binary
+	Lines                         *array.List
+	Line                          *array.Struct
+	LineNumber                    *array.Int64
+	LineFunctionNameIndices       *array.Uint32
+	LineFunctionNameDict          *array.Binary
+	LineFunctionSystemNameIndices *array.Uint32
+	LineFunctionSystemNameDict    *array.Binary
+	LineFunctionFilenameIndices   *array.Uint32
+	LineFunctionFilenameDict      *array.Binary
+	LineFunctionStartLine         *array.Int64
 
 	Value *array.Int64
 	Diff  *array.Int64
@@ -100,47 +100,52 @@ func NewRecordReader(ar arrow.Record) *RecordReader {
 	mappingLimit := location.Field(2).(*array.Uint64)
 	mappingOffset := location.Field(3).(*array.Uint64)
 	mappingFile := location.Field(4).(*array.Dictionary)
+	mappingFileIndices := mappingFile.Indices().(*array.Uint32)
 	mappingFileDict := mappingFile.Dictionary().(*array.Binary)
 	mappingBuildID := location.Field(5).(*array.Dictionary)
+	mappingBuildIDIndices := mappingBuildID.Indices().(*array.Uint32)
 	mappingBuildIDDict := mappingBuildID.Dictionary().(*array.Binary)
 	lines := location.Field(6).(*array.List)
 	line := lines.ListValues().(*array.Struct)
 	lineNumber := line.Field(0).(*array.Int64)
 	lineFunctionName := line.Field(1).(*array.Dictionary)
+	lineFunctionNameIndices := lineFunctionName.Indices().(*array.Uint32)
 	lineFunctionNameDict := lineFunctionName.Dictionary().(*array.Binary)
 	lineFunctionSystemName := line.Field(2).(*array.Dictionary)
+	lineFunctionSystemNameIndices := lineFunctionSystemName.Indices().(*array.Uint32)
 	lineFunctionSystemNameDict := lineFunctionSystemName.Dictionary().(*array.Binary)
 	lineFunctionFilename := line.Field(3).(*array.Dictionary)
+	lineFunctionFilenameIndices := lineFunctionFilename.Indices().(*array.Uint32)
 	lineFunctionFilenameDict := lineFunctionFilename.Dictionary().(*array.Binary)
 	lineFunctionStartLine := line.Field(4).(*array.Int64)
 	valueColumn := ar.Column(labelNum + 1).(*array.Int64)
 	diffColumn := ar.Column(labelNum + 2).(*array.Int64)
 
 	return &RecordReader{
-		Record:                     ar,
-		LabelFields:                labelFields,
-		LabelColumns:               labelColumns,
-		Locations:                  locations,
-		Location:                   location,
-		Address:                    address,
-		MappingStart:               mappingStart,
-		MappingLimit:               mappingLimit,
-		MappingOffset:              mappingOffset,
-		MappingFile:                mappingFile,
-		MappingFileDict:            mappingFileDict,
-		MappingBuildID:             mappingBuildID,
-		MappingBuildIDDict:         mappingBuildIDDict,
-		Lines:                      lines,
-		Line:                       line,
-		LineNumber:                 lineNumber,
-		LineFunctionName:           lineFunctionName,
-		LineFunctionNameDict:       lineFunctionNameDict,
-		LineFunctionSystemName:     lineFunctionSystemName,
-		LineFunctionSystemNameDict: lineFunctionSystemNameDict,
-		LineFunctionFilename:       lineFunctionFilename,
-		LineFunctionFilenameDict:   lineFunctionFilenameDict,
-		LineFunctionStartLine:      lineFunctionStartLine,
-		Value:                      valueColumn,
-		Diff:                       diffColumn,
+		Record:                        ar,
+		LabelFields:                   labelFields,
+		LabelColumns:                  labelColumns,
+		Locations:                     locations,
+		Location:                      location,
+		Address:                       address,
+		MappingStart:                  mappingStart,
+		MappingLimit:                  mappingLimit,
+		MappingOffset:                 mappingOffset,
+		MappingFileIndices:            mappingFileIndices,
+		MappingFileDict:               mappingFileDict,
+		MappingBuildIDIndices:         mappingBuildIDIndices,
+		MappingBuildIDDict:            mappingBuildIDDict,
+		Lines:                         lines,
+		Line:                          line,
+		LineNumber:                    lineNumber,
+		LineFunctionNameIndices:       lineFunctionNameIndices,
+		LineFunctionNameDict:          lineFunctionNameDict,
+		LineFunctionSystemNameIndices: lineFunctionSystemNameIndices,
+		LineFunctionSystemNameDict:    lineFunctionSystemNameDict,
+		LineFunctionFilenameIndices:   lineFunctionFilenameIndices,
+		LineFunctionFilenameDict:      lineFunctionFilenameDict,
+		LineFunctionStartLine:         lineFunctionStartLine,
+		Value:                         valueColumn,
+		Diff:                          diffColumn,
 	}
 }