GetValueIndex performs some work to access the underlying array on every call. Since building the flamegraph accesses the underlying array a lot it makes sense to perform that once, and then access the array.

It turns out that ends up being a lot of work:

```
$ benchstat old.txt new.txt
name                old time/op  new time/op  delta
ArrowFlamegraph-10  1.55ms ± 1%  1.34ms ± 1%  -13.65%  (p=0.000 n=45+10)
```